confluence_generator/requirements.txt
atlassian-python-api
pyyaml
